First try for a connect interceptor
